+++ test/compile.js
@@ -10,13 +10,11 @@ test('compile', (t) => {
 
   t.test('should compile an example code hello world', { timeout: 100000 }, (async (t) => {
     const config = await parse({
-      sourcePath: path.resolve(__dirname, 'fixtures', 'code', '.tryitoutrc.js')
+      sourcePath: path.resolve(__dirname, 'fixtures', 'code', '.tryitout')
     });
 
     compile({
-      config,
-      output: config.output,
-      template: config.template
+      config
     }, (error) => {
       if (error) return t.fail(error);
       t.ok(!fs.existsSync(path.resolve(__dirname, '..', 'docs', 'code', 'build.js')));
@@ -28,13 +26,11 @@ test('compile', (t) => {
 
   t.test('should compile an example product', { timeout: 100000 }, (async (t) => {
     const config = await parse({
-      sourcePath: path.resolve(__dirname, 'fixtures', 'product', '.tryitoutrc.js')
+      sourcePath: path.resolve(__dirname, 'fixtures', 'product', '.tryitout')
     });
 
     compile({
-      config,
-      output: config.output,
-      template: config.template
+      config
     }, (error) => {
       if (error) return t.fail(error);
       t.ok(!fs.existsSync(path.resolve(__dirname, '..', 'docs', 'product', 'build.js')));
@@ -45,13 +41,11 @@ test('compile', (t) => {
 
   t.test('should compile an example landing', { timeout: 100000 }, (async (t) => {
     const config = await parse({
-      sourcePath: path.resolve(__dirname, 'fixtures', 'landing', '.tryitoutrc.js')
+      sourcePath: path.resolve(__dirname, 'fixtures', 'landing', '.tryitout')
     });
 
     compile({
-      config,
-      output: config.output,
-      template: config.template
+      config
     }, (error) => {
       if (error) return t.fail(error);
       t.ok(!fs.existsSync(path.resolve(__dirname, '..', 'docs', 'landing', 'build.js')));
@@ -62,13 +56,11 @@ test('compile', (t) => {
 
   t.test('should compile an example readme', { timeout: 100000 }, (async (t) => {
     const config = await parse({
-      sourcePath: path.resolve(__dirname, 'fixtures', 'readme', '.tryitoutrc.js')
+      sourcePath: path.resolve(__dirname, 'fixtures', 'readme', '.tryitout')
     });
 
     compile({
-      config,
-      output: config.output,
-      template: config.template
+      config
     }, (error) => {
       if (error) return t.fail(error);
       t.ok(!fs.existsSync(path.resolve(__dirname, '..', 'docs', 'readme', 'build.js')));